+++ ext/gettext/gettext.c
@@ -205,7 +205,7 @@ PHP_NAMED_FUNCTION(zif_gettext)
 PHP_NAMED_FUNCTION(zif_dgettext)
 {
 	char *domain, *msgid, *msgstr;
-	int domain_len, msgid_len;
+	size_t domain_len, msgid_len;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ss", &domain, &domain_len, &msgid, &msgid_len) == FAILURE)	{
 		return;
@@ -225,7 +225,7 @@ PHP_NAMED_FUNCTION(zif_dgettext)
 PHP_NAMED_FUNCTION(zif_dcgettext)
 {
 	char *domain, *msgid, *msgstr;
-	int domain_len, msgid_len;
+	size_t domain_len, msgid_len;
 	zend_long category;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ssl", &domain, &domain_len, &msgid, &msgid_len, &category) == FAILURE) {
@@ -246,7 +246,7 @@ PHP_NAMED_FUNCTION(zif_dcgettext)
 PHP_NAMED_FUNCTION(zif_bindtextdomain)
 {
 	char *domain, *dir;
-	int domain_len, dir_len;
+	size_t domain_len, dir_len;
 	char *retval, dir_name[MAXPATHLEN];
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ss", &domain, &domain_len, &dir, &dir_len) == FAILURE) {
@@ -304,7 +304,7 @@ PHP_NAMED_FUNCTION(zif_ngettext)
 PHP_NAMED_FUNCTION(zif_dngettext)
 {
 	char *domain, *msgid1, *msgid2, *msgstr = NULL;
-	int domain_len, msgid1_len, msgid2_len;
+	size_t domain_len, msgid1_len, msgid2_len;
 	zend_long count;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "sssl", &domain, &domain_len,
@@ -330,7 +330,7 @@ PHP_NAMED_FUNCTION(zif_dngettext)
 PHP_NAMED_FUNCTION(zif_dcngettext)
 {
 	char *domain, *msgid1, *msgid2, *msgstr = NULL;
-	int domain_len, msgid1_len, msgid2_len;
+	size_t domain_len, msgid1_len, msgid2_len;
 	zend_long count, category;
 
 	RETVAL_FALSE;
